requirements.txt
click==8.1.3
Pillow==11.3.0
pyserial==3.5
Wand==0.6.10
matplotlib==3.8.4
